グラフ一覧にタブ切替を実装
