+++ components/SiteTopUpper.vue
@@ -1,0 +1,98 @@
+<template>
+  <div>
+    <div class="Header mb-3">
+      <page-header :icon="headerItem.icon">{{ headerItem.title }}</page-header>
+      <div class="UpdatedAt">
+        <span>{{ $t('最終更新') }}</span>
+        <time :datetime="updatedAt">{{ formattedDateForDisplay }}</time>
+      </div>
+      <div
+        v-show="!['ja', 'ja-basic'].includes($i18n.locale)"
+        class="Annotation"
+      >
+        <span>{{ $t('注釈') }}</span>
+      </div>
+    </div>
+    <whats-new class="mb-4" :items="newsItems" :is-emergency="false" />
+    <static-info
+      class="mb-4"
+      :url="localePath('/flow')"
+      :text="$t('自分や家族の症状に不安や心配があればまずは電話相談をどうぞ')"
+      :btn-text="$t('相談の手順を見る')"
+    />
+  </div>
+</template>
+
+<script lang="ts">
+import Vue from 'vue'
+import { MetaInfo } from 'vue-meta'
+import PageHeader from '@/components/PageHeader.vue'
+import WhatsNew from '@/components/WhatsNew.vue'
+import StaticInfo from '@/components/StaticInfo.vue'
+import Data from '@/data/data.json'
+import News from '@/data/news.json'
+import { convertDatetimeToISO8601Format } from '@/utils/formatDate'
+
+export default Vue.extend({
+  components: {
+    PageHeader,
+    WhatsNew,
+    StaticInfo
+  },
+  data() {
+    const data = {
+      Data,
+      headerItem: {
+        icon: 'mdi-chart-timeline-variant',
+        title: this.$t('都内の最新感染動向')
+      },
+      newsItems: News.newsItems
+    }
+    return data
+  },
+  computed: {
+    updatedAt() {
+      return convertDatetimeToISO8601Format(this.$data.Data.lastUpdate)
+    },
+    formattedDateForDisplay() {
+      return this.$d(new Date(Data.lastUpdate), 'dateTime')
+    }
+  },
+  head(): MetaInfo {
+    return {
+      title: this.$t('都内の最新感染動向') as string
+    }
+  }
+})
+</script>
+
+<style lang="scss" scoped>
+.MainPage {
+  .Header {
+    display: flex;
+    flex-wrap: wrap;
+    align-items: flex-end;
+
+    @include lessThan($small) {
+      flex-direction: column;
+      align-items: baseline;
+    }
+  }
+
+  .UpdatedAt {
+    @include font-size(14);
+
+    color: $gray-3;
+    margin-bottom: 0.2rem;
+  }
+
+  .Annotation {
+    @include font-size(12);
+
+    color: $gray-3;
+    @include largerThan($small) {
+      margin: 0 0 0 auto;
+    }
+  }
+}
+</style>
